+++ apps/frontend/src/pages/room/hooks/socket/yjs/canvasSyncHandlers.ts
