+++ apps/frontend/src/pages/room/hooks/socket/yjs/index.ts
@@ -1,0 +1,1 @@
+export { useYjsSocket } from './useYjsSocket'
