+++ apps/frontend/src/pages/room/hooks/socket/yjs/useCanvasTelemetry.ts
